Checks cerbot is installed, add information to configure

+++ certbot.php
@@ -1,17 +1,22 @@
 <?php
 	//Config
 	DEFINE('CERTBOT_PATH',"/root/");
-	if (isset($_POST["createCert"])) {
-		$domain = $_POST["domain"];
-		$acc = $_POST["acc"];
-		$email = $_POST["email"];
-		if (strlen($domain) ==0) {
-			break;
+	if (file_exists(CERTBOT_PATH."/certbot-auto")) {
+		# code...
+		if (isset($_POST["createCert"])) {
+			$domain = $_POST["domain"];
+			$acc = $_POST["acc"];
+			$email = $_POST["email"];
+			if (strlen($domain) ==0) {
+				break;
+			}
+			$command = "cd ".CERTBOT_PATH." && ./certbot-auto certonly --email $email --agree-tos --renew-by-default --webroot  -w /home/$acc/public_html/ -d $domain && cp -f /etc/letsencrypt/live/$domain/fullchain.pem /etc/pki/tls/certs/$domain.crt && cp -f /etc/letsencrypt/live/$domain/privkey.pem /etc/pki/tls/private/$domain.key && cp -f /etc/letsencrypt/live/$domain/chain.pem /etc/pki/tls/certs/$domain.bundle";
+			echo "<pre>";
+			echo shell_exec($command);
+			echo "</pre>";
 		}
-		$command = "cd ".CERTBOT_PATH." && ./certbot-auto certonly --email $email --agree-tos --renew-by-default --webroot  -w /home/$acc/public_html/ -d $domain && cp -f /etc/letsencrypt/live/$domain/fullchain.pem /etc/pki/tls/certs/$domain.crt && cp -f /etc/letsencrypt/live/$domain/privkey.pem /etc/pki/tls/private/$domain.key && cp -f /etc/letsencrypt/live/$domain/chain.pem /etc/pki/tls/certs/$domain.bundle";
-		echo "<pre>";
-		echo shell_exec($command);
-		echo "</pre>";
+	}else{
+		echo "<pre> CERTBOT not installed in '".CERTBOT_PATH."' please configure file</pre>";
 	}
 
 
@@ -19,6 +24,10 @@
  ?>
 	<p>
 		Welcome to the certbot module for CWP
+
+		This relies on CERTBOT already being installed on your system, please configure this file by changing:
+		<code>DEFINE('CERTBOT_PATH',"/root/")</code>
+		To wherever Certbot is installed
 	</p>
  <form method="post">
 	 <label for="domain">Domain Name:</label>
